requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
idna==3.7
influxdb-client==1.43.0
msgpack==1.0.8
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytool==5.0.0
pytz==2024.1
reactivex==4.0.4
requests==2.32.3
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
tinytuya==1.14.0
typing_extensions==4.12.2
urllib3==2.2.1
